feat: Update golangci-lint to version 1.59.0
