+++ deploy/process-compose.prod.yml
@@ -1,4 +1,3 @@
-# Process compose for starting django
 version: "0.5"
 
 processes:
