--- internal/actor/cache.go
@@ -1,1 +1,0 @@
-package actor
