--- internal/local/clients.go
@@ -1,17 +1,0 @@
-package local
-
-import (
-	"github.com/ipfs/kubo/client/rpc"
-)
-
-// IPFSClient is an interface for interacting with an IPFS node.
-type IPFSClient = *rpc.HttpApi
-
-// NewLocalClient creates a new IPFS client that connects to the local IPFS node.
-func GetIPFSClient() (IPFSClient, error) {
-	rpcClient, err := rpc.NewLocalApi()
-	if err != nil {
-		return nil, err
-	}
-	return rpcClient, nil
-}
